refactor: split tauri runtime and review control plane
